+++ libft/ft_substr.c
@@ -6,17 +6,17 @@
 /*   By: cfiliber <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/25 19:59:30 by cfiliber          #+#    #+#             */
-/*   Updated: 2021/11/16 18:12:33 by cfiliber         ###   ########.fr       */
+/*   Updated: 2021/11/25 18:59:26 by cfiliber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_substr(char const *s, unsigned int start, size_t len)
+char	*ft_substr(char const *s, unsigned int start, int len)
 {
 	char	*new_str;
-	size_t	i;
-	size_t	j;
+	int	i;
+	int	j;
 
 	new_str = (char *)malloc(len + 1);
 	if (!s || !new_str)
